backend/requirements.txt
asgiref==3.10.0
Django==5.2.7
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
pillow==12.0.0
sqlparse==0.5.3
tzdata==2025.2
